fix(review): r2 — bounded-poll train tests, raw kernel-validated dg, claim scope
F1 (major): the train-walk tests hid a thread-start race behind a fixed
200 ms sleep (the #178/#176 2-core class — a stalled start sees `done`
first and the drain DISCARDS the queued train). Bounded 10 s poll on
bytes_received before `done`.

F3: the dg clamp invented probe successes GT would not have — GT hands
cJSON's int RAW to setsockopt and lets the kernel EINVAL nonsense
(iperf_udp.c:464-470). set_udp_gso now takes i32 pass-through with
cJSON-style i64 saturation at the server sites; readback pin covers the
round-trip AND the kernel-validation posture (999999/-1 fail; 0 is
kernel-valid GSO-off).

F2: the -P 2 e2e claim downgraded to "exercises" (it cannot
discriminate a pre-loop-only regression) + the sockopt round-trip pin
carries what it can.

F4/F5: probe-vs-reply ordering noted at the accept site (no observable
delta); the deviation note covers short train tails; the len-absent
recompute arm documents GT's differing (equally unreachable) dead arm.

+++ riperf3/src/client.rs
@@ -1326,7 +1326,7 @@ impl Client {
                     // lacking UDP_SEGMENT/UDP_GRO degrades to plain sends
                     // rather than failing the test (#316).
                     if gso_on {
-                        gso_on = net::set_udp_gso(&udp_sock, blksize as u16).is_ok();
+                        gso_on = net::set_udp_gso(&udp_sock, blksize as i32).is_ok();
                     }
                     if gro_on {
                         gro_on = net::set_udp_gro(&udp_sock).is_ok();
